requirements.txt
tweepy
web3
python-dotenv